+++ 02-intro-js/src/data/heroes.js
@@ -1,4 +1,4 @@
-export const heroes = [
+const heroes = [
     {
         id: 1,
         name: 'Batman',
@@ -25,3 +25,14 @@ export const heroes = [
         owner: 'Marvel'
     },
 ];
+
+export const owners = ['Marve', 'DC'];
+
+// We also can export the content of this file to make it default like this
+export default heroes;
+
+// We can export all at once on indivitual exports
+/* export {heroes, owners}; /*/
+
+// We can export all at once and specifying the default export
+/* export { heroes as default, owners}; */
